+++ src/applications/merchant-app/MerchantBuyTab.svelte
@@ -60,7 +60,7 @@
     </h3>
 
     <div class="item-piles-items-list">
-      {#each $itemsPerCategoryStore[category.type] as item (item.id)}
+      {#each $itemsPerCategoryStore[category.type].items as item (item.id)}
         <MerchantItemBuyEntry {item}/>
       {/each}
     </div>